requirements.txt

fastapi
uvicorn[standard]
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
motor
qdrant-client
sentence-transformers
google-generativeai
Pillow
aiofiles
httpx
torch
torchvision
clip-by-openai